+++ pkg/providers/property/testdata/TestResourceEdgeHostname/import_edgehostname_akamaized_product_id.tf
@@ -1,0 +1,11 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_edge_hostname" "importedgehostname" {
+  contract_id   = "ctr_1"
+  group_id      = "grp_2"
+  product_id    = "prd_2"
+  edge_hostname = "test.akamaized.net"
+  ip_behavior   = "IPV4"
+}